+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 
 - Add tooltip when hovering node containing multiple threads
 - Improve node link rendering
-- Display information when the thread graph is not available
+- Display information when the Parallel Stacks is not available
 
 ## 0.1